Store absolute path to files in Structure files
If you run "cover" to create the coverage report in a different directory from where you ran your tests, the coverage report is not able to find the scripts. For large codebases with many scattered tests, it is not feasible to always run the tests from one location. For example, I need coverage data for a script that needs to be run in several different directories.

This change is based on @rolfyone's proposed change in this issue:
    https://github.com/pjcj/Devel--Cover/issues/36